backend/sub_apps/feature_stores/requirements.txt
python-multipart
python-dotenv
fastapi 
uvicorn
sqlalchemy 
psycopg2-binary
pydantic
pandas
numpy 
openpyxl
requests